backend/requirements.txt
asgiref==3.8.1
Django==3.2.3
django-filter==2.4.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
pytz==2025.2
sqlparse==0.5.3
typing_extensions==4.13.2